Update to review comments, mostly documentation and switchin library marker id to usize

+++ datafusion/ffi/src/catalog_provider.rs
@@ -73,7 +73,7 @@ pub struct FFI_CatalogProvider {
 
     /// Utility to identify when FFI objects are accessed locally through
     /// the foreign interface.
-    pub library_marker_id: extern "C" fn() -> u64,
+    pub library_marker_id: extern "C" fn() -> usize,
 }
 
 unsafe impl Send for FFI_CatalogProvider {}
