+++ datafusion/ffi/src/udaf/mod.rs
@@ -148,7 +148,7 @@ pub struct FFI_AggregateUDF {
 
     /// Utility to identify when FFI objects are accessed locally through
     /// the foreign interface.
-    pub library_marker_id: extern "C" fn() -> u64,
+    pub library_marker_id: extern "C" fn() -> usize,
 }
 
 unsafe impl Send for FFI_AggregateUDF {}
