+++ datafusion/ffi/src/udf/mod.rs
@@ -67,8 +67,7 @@ pub struct FFI_ScalarUDF {
     /// FFI equivalent to the `volatility` of a [`ScalarUDF`]
     pub volatility: FFI_Volatility,
 
-    /// Determines the return info of the underlying [`ScalarUDF`]. Either this
-    /// or return_type may be implemented on a UDF.
+    /// Determines the return info of the underlying [`ScalarUDF`].
     pub return_field_from_args: unsafe extern "C" fn(
         udf: &Self,
         args: FFI_ReturnFieldArgs,
@@ -111,7 +110,7 @@ pub struct FFI_ScalarUDF {
 
     /// Utility to identify when FFI objects are accessed locally through
     /// the foreign interface.
-    pub library_marker_id: extern "C" fn() -> u64,
+    pub library_marker_id: extern "C" fn() -> usize,
 }
 
 unsafe impl Send for FFI_ScalarUDF {}
