+++ datafusion/ffi/src/udwf/mod.rs
@@ -108,7 +108,7 @@ pub struct FFI_WindowUDF {
 
     /// Utility to identify when FFI objects are accessed locally through
     /// the foreign interface.
-    pub library_marker_id: extern "C" fn() -> u64,
+    pub library_marker_id: extern "C" fn() -> usize,
 }
 
 unsafe impl Send for FFI_WindowUDF {}
